Add the use of a lockfile
Quote from the Rsnapshot man page: "Lockfile to use when rsnapshot is run. This prevents a second invocation from clobbering the first one. If not specified, no lock file is used. Make sure to use a directory that is not world writeable for security reasons. Use of a lock file is strongly recommended."

Also replaced 'interval' with 'retain'. According to the man page 'interval' is deprecated.

Warning: Changes are not tested. I am new to this all and have yet find out how to build my own test packages. :-D